// src/types/bindings/CreditNoteItemRow.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type CreditNoteItemRow = { id: bigint | null, credit_note_number: string, invoice_item_id: bigint, part_code: string, quantity: number, rate_pre_unit: number, assessable_value: number, cgst_rate: number, cgst_amount: number, sgst_rate: number, sgst_amount: number, igst_rate: number, igst_amount: number, total_value: number, original_quantity: number, original_rate_pre_unit: number, frozen_unit_of_measure: string | null, };